Extract oasysUtils
There was a few utils related to oasys in this file so best to extract
them
